Move metadata yaml from api to core module (#21368)
